web_api/requirements_docker.txt
fastapi
uvicorn
streamlit
opencv-python
numpy
Pillow
redis
requests
